+++ src/components/posts/index.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react'
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles'
 import { grey } from '@material-ui/core/colors'
 import { Box, Divider } from '@material-ui/core'
-import { Post, PostsProps } from 'src/components/posts/post'
+import { Post, PostsProps } from 'src/components/posts/post/index'
 import {
   PostFormSection,
   PostFormSectionProps,
@@ -10,7 +10,7 @@ import {
 import { userHasRoleOnActer } from 'src/lib/user/user-has-role-on-acter'
 import { Acter, ActerConnectionRole, Post as PostType, User } from '@schema'
 
-export interface PostListProps extends PostFormSectionProps {
+export interface PostListProps extends PostFormSectionProps, PostsProps {
   /**
    * Acter on which we are viewing posts
    */
@@ -50,21 +50,23 @@ export const PostList: FC<PostListProps> = ({
       {posts?.map((post) => (
         <Box key={`post-${post.id}`} className={classes.contentContainer}>
           <Post post={post} onPostDelete={onPostDelete} />
-          <Divider variant="middle" className={classes.divider} />
-          {post.Comments?.map((comment) => (
-            <Post
-              key={`post-${post.id}-comment-${comment.id}`}
-              post={comment}
-              commenting
-              onPostDelete={onPostDelete}
-            />
-          ))}
-          <Box>
-            <PostFormSection
-              post={post}
-              user={user}
-              onPostSubmit={onPostSubmit}
-            />
+          <Box className={classes.commentSection}>
+            <Divider className={classes.divider} />
+            {post.Comments?.map((comment) => (
+              <Post
+                key={`post-${post.id}-comment-${comment.id}`}
+                post={comment}
+                commenting
+                onPostDelete={onPostDelete}
+              />
+            ))}
+            <Box>
+              <PostFormSection
+                post={post}
+                user={user}
+                onPostSubmit={onPostSubmit}
+              />
+            </Box>
           </Box>
         </Box>
       ))}
@@ -76,7 +78,7 @@ const useStyles = makeStyles((theme: Theme) =>
   createStyles({
     mainContainer: {
       borderRadius: 7,
-      width: '90%',
+      width: '80%',
       margin: 'auto',
       display: 'flex',
       flexWrap: 'wrap',
@@ -92,6 +94,8 @@ const useStyles = makeStyles((theme: Theme) =>
       display: 'flex',
       flexDirection: 'column',
       padding: theme.spacing(1),
+      paddingLeft: theme.spacing(2.5),
+      paddingRight: theme.spacing(2.5),
       [theme.breakpoints.down('xs')]: {
         width: 300,
       },
@@ -105,6 +109,8 @@ const useStyles = makeStyles((theme: Theme) =>
     },
     commentSection: {
       marginLeft: theme.spacing(6),
+      paddingLeft: theme.spacing(1.5),
+      paddingRight: theme.spacing(1.5),
     },
   })
 )
